minor update to not-found page
